Initialized 2D array and started working on blink.
The 2D array needed to keep track of the all pixels on the board is
initialized and I'll be working with it soon. I've also started working
on the blink method, but I need to find a way to have the savedColor
variable save the value of the color the cursor is currently on, not
the value of the cursor itself.