Blinking cursor and remapped buttons
savedColor variable holds the value of the pixel the cursor is standing
on, and using a counter that increments every time through the loop and
a modulus to check is the counter is even or odd, the cursor now
alternates color between its custom color and the color of the pixel
it's standing on. Also remapped the buttons to change the color of the
cursor to use B instead of A, but I may change that later.