Implementing Suggestions from Classmates
Changed CheckButtonsPress() to CheckButtonsDown() after reading a lot
of people suggest this. It allows the cursor to move and shuffle
through the colors faster. I also decreased the delay and changed the
modulus for blinking the cursor dark to multiples of 4 instead of odd
and even. This makes the window of opportunity for pressing A and
having the cursor stamp a dot greater. It's still not perfect when it
comes to stamping a dot, but the only way I can think of making it
stamp 100% of the time would be getting rid of the blink method, which
is needed to tell what color dot the cursor is standing on.